+++ tests/hom/test_rewriting.cc
@@ -30,9 +30,9 @@ TEST_F(rewriting_test, partition)
                           };
 
     const auto&& p = sdd::hom::rewriter<conf>::partition('a', homs.begin(), homs.end());
-    ASSERT_EQ(1, std::get<0>(p).size()); // F size
-    ASSERT_EQ(1, std::get<1>(p).size()); // G size
-    ASSERT_EQ(1, std::get<2>(p).size()); // L size
+    ASSERT_EQ(static_cast<std::size_t>(1), std::get<0>(p).size()); // F size
+    ASSERT_EQ(static_cast<std::size_t>(1), std::get<1>(p).size()); // G size
+    ASSERT_EQ(static_cast<std::size_t>(1), std::get<2>(p).size()); // L size
     ASSERT_EQ(true, std::get<3>(p)); // has_id
   }
 }
@@ -50,7 +50,7 @@ TEST_F(rewriting_test, sum)
 
     const std::size_t size = cxt.rewrite_cache().size();
     const hom h1 = sdd::hom::rewrite(cxt, h0, 1);
-    ASSERT_EQ(size + 1, cxt.rewrite_cache().size());
+    ASSERT_EQ(static_cast<std::size_t>(size + 1), cxt.rewrite_cache().size());
     ASSERT_NE(h0, h1);
 
     SDD s0(1, SDD(0, {0}, one), SDD(0, {0}, one));
@@ -65,7 +65,7 @@ TEST_F(rewriting_test, sum)
 
     const std::size_t size = cxt.rewrite_cache().size();
     const hom h1 = sdd::hom::rewrite(cxt, h0, 1);
-    ASSERT_EQ(size + 1, cxt.rewrite_cache().size());
+    ASSERT_EQ(static_cast<std::size_t>(size + 1), cxt.rewrite_cache().size());
     ASSERT_NE(h0, h1);
 
     SDD s0(1, SDD(0, {0}, one), SDD(0, {0}, one));
@@ -86,7 +86,7 @@ TEST_F(rewriting_test, fixpoint)
 
     const std::size_t size = cxt.rewrite_cache().size();
     const hom h1 = sdd::hom::rewrite(cxt, h0, 1);
-    ASSERT_EQ(size + 1, cxt.rewrite_cache().size());
+    ASSERT_EQ(static_cast<std::size_t>(size + 1), cxt.rewrite_cache().size());
     ASSERT_NE(h1, h0);
 
     SDD s0(1, SDD(0, {0}, one), SDD(0, {0}, one));
@@ -101,7 +101,7 @@ TEST_F(rewriting_test, fixpoint)
 
     const std::size_t size = cxt.rewrite_cache().size();
     const hom h1 = sdd::hom::rewrite(cxt, h0, 1);
-    ASSERT_EQ(size + 1, cxt.rewrite_cache().size());
+    ASSERT_EQ(static_cast<std::size_t>(size + 1), cxt.rewrite_cache().size());
     ASSERT_EQ(h1, h0);
   }
 }
